fix(navigation): ensure all preceding sections are revealed before scrolling to target hash
