+++ app/utils/sectionNavigation.ts
@@ -27,3 +27,18 @@ export function buildSectionHref(hash: string, isOnHome: boolean, homePath: stri
   if (isOnHome) return targetHash;
   return `${withTrailingSlash(homePath)}${targetHash}`;
 }
+
+export function revealSectionGeometry(targetId?: string) {
+  if (typeof document === 'undefined') return;
+  const sections = [...document.querySelectorAll('.portfolio-content > [id]')] as HTMLElement[];
+  if (!sections.length) return;
+
+  let focusIndex = targetId ? sections.findIndex((section) => section.id === targetId) : -1;
+  if (focusIndex < 0) focusIndex = 0;
+
+  const end = Math.min(sections.length - 1, focusIndex + 1);
+  for (let i = 0; i <= end; i++) {
+    const section = sections[i];
+    if (section) section.style.contentVisibility = 'visible';
+  }
+}
